mini_project/requirements.txt
Flask
flask-cors
google-generativeai
python-dotenv
pywebview
SpeechRecognition